Re factored inventory packets
* moved inventory packets to theirs own folder
* moved inventory handler's to inventory region in ClientPacketHandler
