+++ src/Rasa.Game/Managers/NpcManager.cs
@@ -6,6 +6,7 @@ namespace Rasa.Managers
     using Data;
     using Database.Tables.Character;
     using Game;
+    using Packets.Inventory.Server;
     using Packets.MapChannel.Client;
     using Packets.MapChannel.Server;
     using Structures;
