+++ src/Rasa.Game/Packets/Inventory/Client/ClanLockbox_DestroyItemPacket.cs
@@ -1,11 +1,11 @@
-﻿namespace Rasa.Packets.MapChannel.Client
+﻿namespace Rasa.Packets.Inventory.Client
 {
     using Data;
     using Memory;
 
-    public class PersonalInventory_DestroyItemPacket : ClientPythonPacket
+    public class ClanLockbox_DestroyItemPacket : ClientPythonPacket
     {
-        public override GameOpcode Opcode { get; } = GameOpcode.PersonalInventory_DestroyItem;
+        public override GameOpcode Opcode { get; } = GameOpcode.ClanLockbox_DestroyItem;
 
         public ulong EntityId { get; set; }
         public uint Quantity { get; set; }
