+++ src/Rasa.Game/Packets/Inventory/Client/WeaponDrawerInventory_MoveItemPacket.cs
@@ -1,21 +1,19 @@
-﻿namespace Rasa.Packets.MapChannel.Client
+﻿namespace Rasa.Packets.Inventory.Client
 {
     using Data;
     using Memory;
 
-    public class RequestEquipWeaponPacket : ClientPythonPacket
+    public class WeaponDrawerInventory_MoveItemPacket : ClientPythonPacket
     {
-        public override GameOpcode Opcode { get; } = GameOpcode.RequestEquipWeapon;
-
+        public override GameOpcode Opcode { get; } = GameOpcode.WeaponDrawerInventory_MoveItem;
+        
         public uint SrcSlot { get; set; }
-        public InventoryType InventoryType { get; set; }
         public uint DestSlot { get; set; }
 
         public override void Read(PythonReader pr)
         {
             pr.ReadTuple();
             SrcSlot = pr.ReadUInt();
-            InventoryType = (InventoryType)pr.ReadInt();
             DestSlot = pr.ReadUInt();
         }
     }
