+++ src/Rasa.Game/Packets/Inventory/Server/AddBuybackItemPacket.cs
@@ -1,4 +1,4 @@
-﻿namespace Rasa.Packets.MapChannel.Server
+﻿namespace Rasa.Packets.Inventory.Server
 {
     using Data;
     using Memory;
